impl faucet ffi
