+++ faucet/ffi/Cargo.toml
@@ -1,7 +1,15 @@
 [package]
 name = "sunshine-faucet-ffi"
 version = "0.1.0"
-authors = ["David Craven <david@craven.ch>"]
+authors = ["David Craven <david@craven.ch>", "Shady Khalifa <shekohex@gmail.com>"]
 edition = "2018"
 
 [dependencies]
+sunshine-ffi-utils = { git = "https://github.com/sunshine-protocol/sunshine-core" }
+sunshine-client-utils = { git = "https://github.com/sunshine-protocol/sunshine-core" }
+substrate-subxt = "0.11.0"
+sunshine-faucet-client = { version = "^0.2", path = "../../faucet/client" }
+thiserror = "1.0.20"
+
+[dev-dependencies]
+test-client = { path = "../../bin/client" }
