+++ identity/ffi/Cargo.toml
@@ -11,7 +11,7 @@ repository = "https://github.com/sunshine-protocol/sunshine-identity"
 [dependencies]
 substrate-subxt = "0.11.0"
 sunshine-client-utils = { git = "https://github.com/sunshine-protocol/sunshine-core" }
-sunshine-faucet-client = { version = "^0.2", path = "../../faucet/client", optional = true }
+sunshine-faucet-ffi = { version = "^0.1", path = "../../faucet/ffi", optional = true }
 sunshine-ffi-utils = { git = "https://github.com/sunshine-protocol/sunshine-core" }
 sunshine-identity-client = { version = "^0.2", path = "../client" }
 thiserror = "1.0.20"
@@ -21,4 +21,4 @@ test-client = { path = "../../bin/client" }
 
 [features]
 default = []
-faucet = ["sunshine-faucet-client"]
+faucet = ["sunshine-faucet-ffi"]
