+++ identity/ffi/src/lib.rs
@@ -1,4 +1,7 @@
 pub use sunshine_ffi_utils as utils;
+
+#[cfg(feature = "faucet")]
+pub use sunshine_faucet_ffi as faucet_ffi;
 #[doc(hidden)]
 pub mod ffi;
 
@@ -81,13 +84,9 @@ macro_rules! impl_ffi {
                 identifier: *const raw::c_char = cstr!(identifier),
                 amount: u64 = amount
             ) -> String;
-
-            /// Try to mint the current account, this only enabled in testnet and behind a feature flag
-            /// returned the minted amount or null if there is any errors
-            #[cfg(feature = "faucet")]
-            Faucet::mint => fn client_faucet_mint() -> String;
-
-        }
+        };
+        #[cfg(feature = "faucet")]
+        $crate::faucet_ffi::impl_ffi!();
     };
     (client: $client: ty) => {
         gen_ffi!(client = $client);
